Create README.md
Computer vision project implemented with OpenCV

Ever wanted to draw your imagination by just waving your finger in the air. In this post, we will learn to build an Air Canvas which can draw anything on it by just capturing the motion of a coloured marker with a camera. Here a coloured object at the tip of the finger is used as the marker.

We will be using the computer vision techniques of OpenCV to build this project. The preferred language is python due to its exhaustive libraries and easy to use syntax but understanding the basics can be implemented in any OpenCV supported language.

Here Colour Detection and tracking are used in order to achieve the objective. The colour marker is detected and a mask is produced. It includes the further steps of morphological operations on the mask produced which are Erosion and Dilation. Erosion reduces the impurities present in the mask and dilation further restores the eroded main mask.